third_party/blink/web_tests/external/wpt/css/css-text/text-transform/math/text-transform-math-auto-003-expected.txt

This is a testharness.js-based test.
Found 111 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Selection.toString() for math-auto 'A' returns the transformed character.
  assert_equals: expected "𝐴" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'B' returns the transformed character.
  assert_equals: expected "𝐵" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'C' returns the transformed character.
  assert_equals: expected "𝐶" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'D' returns the transformed character.
  assert_equals: expected "𝐷" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'E' returns the transformed character.
  assert_equals: expected "𝐸" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'F' returns the transformed character.
  assert_equals: expected "𝐹" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'G' returns the transformed character.
  assert_equals: expected "𝐺" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'H' returns the transformed character.
  assert_equals: expected "𝐻" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'I' returns the transformed character.
  assert_equals: expected "𝐼" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'J' returns the transformed character.
  assert_equals: expected "𝐽" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'K' returns the transformed character.
  assert_equals: expected "𝐾" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'L' returns the transformed character.
  assert_equals: expected "𝐿" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'M' returns the transformed character.
  assert_equals: expected "𝑀" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'N' returns the transformed character.
  assert_equals: expected "𝑁" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'O' returns the transformed character.
  assert_equals: expected "𝑂" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'P' returns the transformed character.
  assert_equals: expected "𝑃" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Q' returns the transformed character.
  assert_equals: expected "𝑄" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'R' returns the transformed character.
  assert_equals: expected "𝑅" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'S' returns the transformed character.
  assert_equals: expected "𝑆" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'T' returns the transformed character.
  assert_equals: expected "𝑇" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'U' returns the transformed character.
  assert_equals: expected "𝑈" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'V' returns the transformed character.
  assert_equals: expected "𝑉" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'W' returns the transformed character.
  assert_equals: expected "𝑊" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'X' returns the transformed character.
  assert_equals: expected "𝑋" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Y' returns the transformed character.
  assert_equals: expected "𝑌" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Z' returns the transformed character.
  assert_equals: expected "𝑍" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'a' returns the transformed character.
  assert_equals: expected "𝑎" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'b' returns the transformed character.
  assert_equals: expected "𝑏" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'c' returns the transformed character.
  assert_equals: expected "𝑐" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'd' returns the transformed character.
  assert_equals: expected "𝑑" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'e' returns the transformed character.
  assert_equals: expected "𝑒" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'f' returns the transformed character.
  assert_equals: expected "𝑓" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'g' returns the transformed character.
  assert_equals: expected "𝑔" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'i' returns the transformed character.
  assert_equals: expected "𝑖" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'j' returns the transformed character.
  assert_equals: expected "𝑗" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'k' returns the transformed character.
  assert_equals: expected "𝑘" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'l' returns the transformed character.
  assert_equals: expected "𝑙" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'm' returns the transformed character.
  assert_equals: expected "𝑚" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'n' returns the transformed character.
  assert_equals: expected "𝑛" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'o' returns the transformed character.
  assert_equals: expected "𝑜" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'p' returns the transformed character.
  assert_equals: expected "𝑝" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'q' returns the transformed character.
  assert_equals: expected "𝑞" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'r' returns the transformed character.
  assert_equals: expected "𝑟" but got "U+d835"
[FAIL] Selection.toString() for math-auto 's' returns the transformed character.
  assert_equals: expected "𝑠" but got "U+d835"
[FAIL] Selection.toString() for math-auto 't' returns the transformed character.
  assert_equals: expected "𝑡" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'u' returns the transformed character.
  assert_equals: expected "𝑢" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'v' returns the transformed character.
  assert_equals: expected "𝑣" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'w' returns the transformed character.
  assert_equals: expected "𝑤" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'x' returns the transformed character.
  assert_equals: expected "𝑥" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'y' returns the transformed character.
  assert_equals: expected "𝑦" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'z' returns the transformed character.
  assert_equals: expected "𝑧" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ı' returns the transformed character.
  assert_equals: expected "𝚤" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ȷ' returns the transformed character.
  assert_equals: expected "𝚥" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Α' returns the transformed character.
  assert_equals: expected "𝛢" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Β' returns the transformed character.
  assert_equals: expected "𝛣" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Γ' returns the transformed character.
  assert_equals: expected "𝛤" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Δ' returns the transformed character.
  assert_equals: expected "𝛥" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Ε' returns the transformed character.
  assert_equals: expected "𝛦" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Ζ' returns the transformed character.
  assert_equals: expected "𝛧" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Η' returns the transformed character.
  assert_equals: expected "𝛨" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Θ' returns the transformed character.
  assert_equals: expected "𝛩" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Ι' returns the transformed character.
  assert_equals: expected "𝛪" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Κ' returns the transformed character.
  assert_equals: expected "𝛫" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Λ' returns the transformed character.
  assert_equals: expected "𝛬" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Μ' returns the transformed character.
  assert_equals: expected "𝛭" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Ν' returns the transformed character.
  assert_equals: expected "𝛮" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Ξ' returns the transformed character.
  assert_equals: expected "𝛯" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Ο' returns the transformed character.
  assert_equals: expected "𝛰" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Π' returns the transformed character.
  assert_equals: expected "𝛱" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Ρ' returns the transformed character.
  assert_equals: expected "𝛲" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Σ' returns the transformed character.
  assert_equals: expected "𝛴" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Τ' returns the transformed character.
  assert_equals: expected "𝛵" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Υ' returns the transformed character.
  assert_equals: expected "𝛶" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Φ' returns the transformed character.
  assert_equals: expected "𝛷" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Χ' returns the transformed character.
  assert_equals: expected "𝛸" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Ψ' returns the transformed character.
  assert_equals: expected "𝛹" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'Ω' returns the transformed character.
  assert_equals: expected "𝛺" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'α' returns the transformed character.
  assert_equals: expected "𝛼" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'β' returns the transformed character.
  assert_equals: expected "𝛽" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'γ' returns the transformed character.
  assert_equals: expected "𝛾" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'δ' returns the transformed character.
  assert_equals: expected "𝛿" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ε' returns the transformed character.
  assert_equals: expected "𝜀" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ζ' returns the transformed character.
  assert_equals: expected "𝜁" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'η' returns the transformed character.
  assert_equals: expected "𝜂" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'θ' returns the transformed character.
  assert_equals: expected "𝜃" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ι' returns the transformed character.
  assert_equals: expected "𝜄" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'κ' returns the transformed character.
  assert_equals: expected "𝜅" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'λ' returns the transformed character.
  assert_equals: expected "𝜆" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'μ' returns the transformed character.
  assert_equals: expected "𝜇" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ν' returns the transformed character.
  assert_equals: expected "𝜈" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ξ' returns the transformed character.
  assert_equals: expected "𝜉" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ο' returns the transformed character.
  assert_equals: expected "𝜊" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'π' returns the transformed character.
  assert_equals: expected "𝜋" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ρ' returns the transformed character.
  assert_equals: expected "𝜌" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ς' returns the transformed character.
  assert_equals: expected "𝜍" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'σ' returns the transformed character.
  assert_equals: expected "𝜎" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'τ' returns the transformed character.
  assert_equals: expected "𝜏" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'υ' returns the transformed character.
  assert_equals: expected "𝜐" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'φ' returns the transformed character.
  assert_equals: expected "𝜑" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'χ' returns the transformed character.
  assert_equals: expected "𝜒" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ψ' returns the transformed character.
  assert_equals: expected "𝜓" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ω' returns the transformed character.
  assert_equals: expected "𝜔" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ϑ' returns the transformed character.
  assert_equals: expected "𝜗" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ϕ' returns the transformed character.
  assert_equals: expected "𝜙" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ϖ' returns the transformed character.
  assert_equals: expected "𝜛" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ϰ' returns the transformed character.
  assert_equals: expected "𝜘" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ϱ' returns the transformed character.
  assert_equals: expected "𝜚" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ϴ' returns the transformed character.
  assert_equals: expected "𝛳" but got "U+d835"
[FAIL] Selection.toString() for math-auto 'ϵ' returns the transformed character.
  assert_equals: expected "𝜖" but got "U+d835"
[FAIL] Selection.toString() for math-auto '∂' returns the transformed character.
  assert_equals: expected "𝜕" but got "U+d835"
[FAIL] Selection.toString() for math-auto '∇' returns the transformed character.
  assert_equals: expected "𝛻" but got "U+d835"
Harness: the test ran to completion.
